# Implement layout="zip" for Lambda/GCF, deprecating lambdex