--- a/src/python/pants/backend/awslambda/python/rules_test.py
+++ b/src/python/pants/backend/awslambda/python/rules_test.py
@@ -294,7 +294,7 @@ def test_create_hello_world_lambda(rule_runner: PythonRuleRunner) -> None:
     zip_file_relpath, content = create_python_awslambda(
         rule_runner,
         Address("src/python/foo/bar", target_name="lambda"),
-        expected_extra_log_lines=("    Handler: foo.bar.hello_world.handler",),
+        expected_extra_log_lines=("    Handler: lambda_function.handler",),
         extra_args=["--lambdex-layout=zip"],
     )
     assert "src.python.foo.bar/lambda.zip" == zip_file_relpath
@@ -303,11 +303,14 @@ def test_create_hello_world_lambda(rule_runner: PythonRuleRunner) -> None:
     names = set(zipfile.namelist())
     assert "mureq/__init__.py" in names
     assert "foo/bar/hello_world.py" in names
+    assert (
+        zipfile.read("lambda_function.py") == b"from foo.bar.hello_world import handler as handler"
+    )
 
     zip_file_relpath, content = create_python_awslambda(
         rule_runner,
         Address("src/python/foo/bar", target_name="slimlambda"),
-        expected_extra_log_lines=("    Handler: foo.bar.hello_world.handler",),
+        expected_extra_log_lines=("    Handler: lambda_function.handler",),
         extra_args=["--lambdex-layout=zip"],
     )
     assert "src.python.foo.bar/slimlambda.zip" == zip_file_relpath
@@ -316,3 +319,6 @@ def test_create_hello_world_lambda(rule_runner: PythonRuleRunner) -> None:
     names = set(zipfile.namelist())
     assert "mureq/__init__.py" not in names
     assert "foo/bar/hello_world.py" in names
+    assert (
+        zipfile.read("lambda_function.py") == b"from foo.bar.hello_world import handler as handler"
+    )